XAP-12508,XAP-13033
changed method name, added unit test

+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/cluster/node/impl/backlog/BacklogWeightPolicy.java
@@ -8,5 +8,5 @@ import com.gigaspaces.internal.cluster.node.impl.packets.data.IReplicationPacket
  */
 public interface BacklogWeightPolicy {
 
-    int calculateWeight(IReplicationPacketData<?> data);
+    long calculateWeight(IReplicationPacketData<?> data);
 }
